dev-perl/Tk: allow compilation while using deprecated function
the function XKeycodeToKeysym has been deprecated, and its signature has
changed. while this compiles "fine" on gcc, it doesn't on clang-16.

ideally, we should actually remove all references to this function, but
to allow it to actually compile on >=clang-16 i'm just casting it to the
expected types instead.

this is just a quick and dirty fix.

Signed-off-by: Mazunki Hoksaas <rolferen@gmail.com>